Bind the two upload-event match scrutinees to a name and take clippy's mechanical fixes across the tree
`cargo clippy --fix` over lib, bins and tests. Everything in it is a
machine-applied suggestion with no behaviour change — redundant clones on
`Copy` types, `&vec.clone()` where a slice reference does, no-op casts,
`or_insert_with(Default::default)`, identity `map`, `format!` with nothing to
interpolate, `let ... else` collapsed to `?`, and similar. 157 warnings down to
88.

The automatic pass could not run until the two `UploadEventKind` arms in
`network/mod.rs` were fixed by hand. Both scrutinised a block directly —
`match { let mut mgr = ...write().await; ...; promoted } { ... }` — and
`blocks_in_conditions` rewrites that into `let promoted = let res = {`, which
is not an expression. Cargo compiles the whole crate after applying fixes and
rolls the entire run back when it fails, so those two sites were quietly
costing every other fix in the tree. They now bind the block to a name and
match on that, which also makes it obvious that the write guard is released
before the arm body runs rather than being held across it.

The one large diff is `ed2k/friend_connect.rs`, where `items_after_test_module`
moved 504 lines of packet helpers from below `#[cfg(test)] mod tests` to above
it. Item order carries no meaning in Rust so the move is inert, but the file
reads as almost entirely rewritten when it is the same code in a different
place — `git blame -M` follows it.

Verified: cargo test --lib 1598 passed, cargo check clean. The remaining 88 are
the ones clippy will not fix on its own and are left for a manual pass.

+++ src-tauri/src/network/ed2k/aich.rs
@@ -153,7 +153,7 @@ pub(crate) fn hierarchical_root(all_leaves: &[[u8; 20]], file_size: u64) -> [u8;
         return all_leaves[0];
     }
 
-    let num_parts = ((file_size as u64 + PARTSIZE as u64 - 1) / PARTSIZE as u64) as usize;
+    let num_parts = ((file_size + PARTSIZE as u64 - 1) / PARTSIZE as u64) as usize;
     if num_parts <= 1 {
         return build_tree_recursive(all_leaves, true);
     }
@@ -978,7 +978,7 @@ mod tests {
     fn corrupt_blocks_from_aich_recovery_roundtrip() {
         let data = vec![0x42u8; AICH_BLOCK_SIZE * 2];
         let trusted = AICHRecoveryHashSet::build_from_data(&data);
-        let recovery = trusted.create_part_recovery_data(0, PARTSIZE as usize);
+        let recovery = trusted.create_part_recovery_data(0, PARTSIZE);
         let mut bad = data.clone();
         bad[AICH_BLOCK_SIZE + 10] ^= 0xFF;
         let corrupt = corrupt_blocks_from_aich_recovery(
